TraceTarget no longer blocks on error messages
When using TraceTarget to log error events, it would cause the application
to block because of the default behaviour of Trace.Fail (described on the
MSDN page).
Changed to only use Trace.Fail when receiving an event Fatal or above.
Also started to use TraceInformation, TraceWarning and TraceError for the
correspondig log levels.