BUG 6737 UserStackFrame wrong when logger wrapped
When using a wrapper class around the logger, such as in
Common.Logging, the UserStackFrame would point to the wrapper log method,
instead of the user method.
The solution implemented is to only skip an assembly if loggerType is
null, which will use the first non NLog/mscorlib/System call.
Thanks to @aloker for providing the patch.