Add test for race condition in async flushing
To problem occurs when the target is flushed twice, before the first flush
has been started. The reason is that all the continuations aren't called,
normally this wouldn't be a problem. But if configured with a SplitTarget
splitting to the async target and calling LogManager.Flush(), all targets
will flush, meaning the async target will be flused twice in a row,
resulting in a timeout.