Use merging of event properties in async target wrapper to fix empty collection issue
The empty 'properties' collection (issue #331) was fixed by copying the properties from the original object stored in the parameters. However, the async target wrapper 'Write' is overwritten such that the merge method is never called, causing the same issue when using <targets async="true"> configuration. The solution is to call the merge operation in the overridden 'Write' method.