LogFactory Class Refactored
* New inner class LoggerCache added. It encapsulates the operations
around Dictionary loggerCache
* The LoggerCacheKey class now implements the IEquatable. This can
potentially help performance in cases it is invoked
* Removed ReconfigExistingLoggers(LoggingConfiguration) internal method.
ReconfigExistingLoggers(LoggingConfiguration) was merged into
ReconfigExistingLoggers() public method. This is because the internal
one was always invoked with parameter value of this.config
* Renamed DisableLogging() and EnableLogging() methods in LogManager
class to SuspendLogging() and ResumeLogging() respectively. This will
align the naming with Microsoft conventions (see Control.ResumeLayout
method). The DisableLogging() and EnableLogging() were marked as
obsolete.
* Remove CLSCompliant(false) attribute from various methods as it is not
required.
* Comments and code reformatted.
* [Fix] Narrowed the captured exceptions within
InvalidLoggerConfiguration tests in UnitTests\GetLoggerTests class.

[**Note**] The name changes on DisableLogging() and EnableLogging()
methods affect the public interface.