Added code to preserve alignment when truncating fixed length layouts.
Updated PaddingLayoutRenderWrapper.cs to remove characters from the start of the string instead of the end when Padding is positive. (The previous behaviour of removing characters from the end is still used when Padding is negative.)
Added test file LayoutRenderers/Wrappers/PaddingTests.cs with test cases to ensure the alignment behaviour of fixed-length padding is correct. Added PaddingTests.cs to each of the NLog.UnitTests.*.csproj files.