Added a Base Class for Wcf Log Receivers
The base class allows for the code duplication to removed.  Also, it
allows for the inheriting class to invoke the correct channel from the
ClientBase<TService>, so WCF should be happy at runtime, too.

Added deprecation to ILogReceiverClient and its implementation in favor
of the other two interfaces.

Added deprecation to the LogReceiverWebServiceTarget's protected method
that returned a concrete class in favor of a new method that returns an
interface.