Major pooling effort to minimize garbage collections when running logging. LogEventInfo gets pooled now if configured via config file, plus a lot of other objects that is used for getting the log event into the resulting file. Changed of some Inline Lambda functions into classes that gets pooled, to not allocate the auto generated class that gets allocated every time a lambda function gets called.
Fixed "bug" in AsyncTargetWrapper.FlushAsync, since it allowed flushes to go on and not wait for it, which is definately one reason why the unit tests that uses the async target wrapper is flaky.
Made File Comparison more rubust in the sense that FileInfo can cache information, which can mean that when accessing properties, its not always up to date with the underlying file system.
Added Sync method on Mono for tests, so that filesystem is forced to flush to disk