Refactor InternalLogger class
Refactor InternalLogger class to improve code readabiity and create further opportunities to improve the code.

There are no functional changes or changes to the public interface.

The method Write() in InternalLogger class split into smaller methods
* FormatMessage()
* WriteToLogFile()
* WriteToTextWriter()
* WriteToConsole()
* WriteToErrorConsole()

LoggingEnabled() method renamed to IsLoggingEnabled() in alignment to IsSeriousException(). It also convey better to the reader the purpose of the method.